+++ webview-ui/components/add-tab-dropdown.ts
@@ -105,4 +105,3 @@ declare global {
         'add-tab-dropdown': AddTabDropdown;
     }
 }
-