+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to the "PengSheets" extension will be documented in this fil
 
 ### Fixed
 - Fix context menu and add-tab dropdown being clipped when displayed near viewport edges.
+- Fix + button "Add New Document" inserting document at wrong position in hybrid notebooks.
+
 ## [1.0.4] - 2026-01-10
 
 ### Fixed